requirements.txt
# requirements.txt
streamlit==1.32.2
google-auth==2.23.4
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas==2.1.4
httplib2==0.22.0
python-dateutil==2.8.2
